Function annotations for end-user backtracing
This patch provides a couple of extra options to the uglifyjs command
line, that allow you to provide each function with run-time accessible
information about the function's original line number (and source file).

Also an option is added to start each function body on a new line in the
output file, such that even browsers with very poor error reporting will
be able to derive (a backtrace of) source level functions.

These options combined, the output will look similar to:

foo();bar(function(){
3|14};function
x(){3|18};

The interesting part is the <fileNum>|<lineNum> at the start of each
function. Each file is assigned a static number in the json file
maintained by the --annotation-map option, to keep the amount of added
bytes to a minimum.

Using Function.prototype.toString() and a simple RegExp, we can access
this information at run time when we have a reference to a function. In
case the browser just gave us a line number, we can download the
offending javascript file using an XMLHttpRequest, look up the offending
line, and do the same.

The multi-file support assumes that the single input file has already
been combined by some other preprocessor (or the a simple script doing
some cat'ting), that adds lines of this form:

//@@path/to/file.coffee:0

The '0' is the line number that the comment line would have in the file,
so in this example the next line would be '1'.